requirements.txt
# Discord Bot Dependencies
discord.py>=2.0.0
aiohttp>=3.8.0
python-dotenv>=0.19.0

# Database Dependencies
SQLAlchemy>=1.4.0
psycopg2-binary>=2.9.0
alembic==1.12.1

# API Dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2

# Utility Dependencies
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1

# Testing Dependencies
pytest==7.4.3
pytest-asyncio==0.21.1

# Development Dependencies
black==23.11.0
flake8==6.1.0
isort==5.12.0

# Google API Dependencies
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.4.0